Refactor of plotUtils to place component specific logic into the component it belongs too. Proper SoC between plot and utils now. Individual plots need to be optimized further.
